operator: add AdditionalInitContainers, HostNetwork, and DnsPolicy options to ProxyClass
This adds a method for ProxyClasses to perform additional setup for specific
environment needs, such as updating a network routing rule or to set additional
parameters with sysctl before the proxy starts. Additionally, exposing HostNetwork
and DnsPolicy -- usually the latter needs to be changed when enabling hostNetwork -- lets
the Proxy pod configure routing logic on underlying machine, if needed for your environment.